requirements.txt
Flask
Selenium
beautifulsoup4
requests
python-dotenv